Move CodeLens class and instances into CodeLens handler
