+++ src/Curry/LanguageServer/Handlers.hs
@@ -12,7 +12,6 @@ import Curry.LanguageServer.Handlers.TextDocument (didOpenHandler, didChangeHand
 import Curry.LanguageServer.Handlers.WorkspaceSymbols (workspaceSymbolHandler)
 import Curry.LanguageServer.Monad (LSM)
 import qualified Language.LSP.Server as S
-import qualified Language.LSP.Types as J
 
 handlers :: S.Handlers LSM
 handlers = mconcat
@@ -30,4 +29,5 @@ handlers = mconcat
     , didOpenHandler
     , didChangeHandler
     , didSaveHandler
+    , didCloseHandler
     ]
